Fix critical issues and update project configuration
- Fix GPS coordinates: Change Double to Float in Post model
- Add bird name validation (letters only) in AddPost and EditPost fragments
- Remove Room/KAPT dependencies to resolve compilation errors
- Update Gradle configuration and SDK versions (compileSdk 36)
- Fix launcher icons and app configuration
- Remove problematic ui/theme files causing compiler issues

Ready for maps integration and navigation implementation

+++ app/src/main/java/model/Post.kt
@@ -9,8 +9,8 @@ data class Post(
     var scientificName: String = "",
     var description: String = "",
     var imageUrl: String = "",
-    var latitude: Double = 0.0,
-    var longitude: Double = 0.0,
+    var latitude: Float = 0.0f,
+    var longitude: Float = 0.0f,
     var location: String = "",
     var timestamp: Long = 0L
 )
